refactor(breakdown): avoid circular imports
