+++ dessia_common/breakdown.py
@@ -8,9 +8,7 @@ import collections
 import collections.abc
 import numpy as npy
 
-from dessia_common import REF_MARKER, OLD_REF_MARKER
 import dessia_common.serialization as dcs
-import dessia_common.utils.types as dct
 
 
 def attrmethod_getter(object_, attr_methods):
@@ -31,96 +29,6 @@ def attrmethod_getter(object_, attr_methods):
         else:
             object_ = getattr(object_, segment)
     return object_
-
-
-class ExtractionError(Exception):
-    """ Custom Exception for deep attributes Extraction process. """
-
-
-def extract_segment_from_object(object_, segment: str):
-    """ Try all ways to get an attribute (segment) from an object that can of numerous types. """
-    if dct.is_sequence(object_):
-        try:
-            return object_[int(segment)]
-        except ValueError as err:
-            message_error = (f"Cannot extract segment {segment} from object {{str(object_)[:500]}}:"
-                             f" segment is not a sequence index")
-            raise ExtractionError(message_error) from err
-
-    if isinstance(object_, dict):
-        if segment in object_:
-            return object_[segment]
-
-        if segment.isdigit():
-            intifyed_segment = int(segment)
-            if intifyed_segment in object_:
-                return object_[intifyed_segment]
-            if segment in object_:
-                return object_[segment]
-            raise ExtractionError(f'Cannot extract segment {segment} from object {str(object_)[:200]}')
-
-        # should be a tuple
-        if segment.startswith('(') and segment.endswith(')') and ',' in segment:
-            key = []
-            for subsegment in segment.strip('()').replace(' ', '').split(','):
-                if subsegment.isdigit():
-                    subkey = int(subsegment)
-                else:
-                    subkey = subsegment
-                key.append(subkey)
-            return object_[tuple(key)]
-        raise ExtractionError(f"Cannot extract segment {segment} from object {str(object_)[:500]}")
-
-    # Finally, it is a regular object
-    return getattr(object_, segment)
-
-
-def get_in_object_from_path(object_, path, evaluate_pointers=True):
-    """ Get deep attributes from an object. Argument 'path' represents path to deep attribute. """
-    segments = path.lstrip('#/').split('/')
-    element = object_
-    for segment in segments:
-        if isinstance(element, dict):
-            # Going down in the object and it is a reference : evaluating sub-reference
-            if evaluate_pointers:
-                if REF_MARKER in element:
-                    try:
-                        element = get_in_object_from_path(object_, element[REF_MARKER])
-                    except RecursionError as err:
-                        err_msg = f'Cannot get segment {segment} from path {path} in element {str(element)[:500]}'
-                        raise RecursionError(err_msg) from err
-                elif OLD_REF_MARKER in element:  # Retro-compatibility to be remove sometime
-                    try:
-                        element = get_in_object_from_path(object_, element[OLD_REF_MARKER])
-                    except RecursionError as err:
-                        err_msg = f'Cannot get segment {segment} from path {path} in element {str(element)[:500]}'
-                        raise RecursionError(err_msg) from err
-
-        try:
-            element = extract_segment_from_object(element, segment)
-        except ExtractionError as err:
-
-            err_msg = f'Cannot get segment {segment} from path {path} in element {str(element)[:500]}'
-            raise ExtractionError(err_msg) from err
-
-    return element
-
-
-def set_in_object_from_path(object_, path, value, evaluate_pointers=True):
-    """ Set deep attribute from an object to the given value. Argument 'path' represents path to deep attribute. """
-    reduced_path = '/'.join(path.lstrip('#/').split('/')[:-1])
-    last_segment = path.split('/')[-1]
-    if reduced_path:
-        last_object = get_in_object_from_path(object_, reduced_path, evaluate_pointers=evaluate_pointers)
-    else:
-        last_object = object_
-
-    if dct.is_sequence(last_object):
-        last_object[int(last_segment)] = value
-    elif isinstance(last_object, dict):
-        last_object[last_segment] = value
-    else:
-        setattr(last_object, last_segment, value)
 
 
 def merge_breakdown_dicts(dict1, dict2):
